control-cd/requirements.txt
Flask>=2.1.0
Werkzeug>=2.1.0
GitPython==3.1.14
